feat(utils): Created an specific FSM subClass for Number
